backend/auth-service/requirements.txt
Django>=3.0
psycopg2>2.8
djangorestframework>3.0
PyJWT>=2.8
djangorestframework-simplejwt>5.0

